Move elm-package.json into test/
Use the ability to specify cwd to put elm-package.json in with the test files.

This also uses path.resolve() on the file paths, because we can no longer rely
on them being the correct relative paths once they're passed to elm-make.  This
more closely reflects the sort of file we get through gulp, anyway.