Don't error when compiler doesn't generate output
If an elm file is passed into the 0.19 compiler and it doesn't have a main
entrypoint, no file is generated.  Since this is expected behavior, gulp-elm no
longer throws an Error here (assuming there will be an error during the compile
stage if the file isn't generated because of a problem).

Fixes #37